Update Waveform_Camera_Sync_Acquisition.m
Previously, the use of the camera as the clock source was hardcoded into this script, bypassing any settings to the contrary on the waveforms tab. Now, setting the clock to internal on the waveforms tab will allow internally-clocked acquisition using this script, with clock frequency determined by the setting on the waveforms tab.